Remove some todo!(), unwrap, run clippy in CI (#93)
* Add new testing

* refactor

* remove todos

* remove todos

* remove todos

* remove todos

* fix import star

* before clippy

* before clippy

* before clippy

* before clippy

* before clippy

* clippy

* clippy

* st mapping proj

* st indexed access

* clippy

* clippy

* pre clippy

* clippy

* clippy

+++ packages/beff-core/src/ast/json.rs
@@ -63,7 +63,7 @@ impl Json {
         } else if let Some(it) = it.as_u64() {
             Self::parse_int(it as i64)
         } else if let Some(it) = it.as_i64() {
-            Self::parse_int(it as i64)
+            Self::parse_int(it)
         } else {
             unreachable!("should be possible to convert serde_json::Number to Json::Number")
         }
